requirements.txt
fastparquet==2023.7.0
pandas==2.0.3
pandas-gbq==0.19.2
prefect==2.10.21
prefect-gcp==0.4.4
prefect_gcp[bigquery]==0.4.4
pyarrow==12.0.1
python-dotenv==1.0.0
selenium==4.10.0
Scrapy==2.9.0
webdriver-manager==3.8.6
